+++ src/posting/widgets/request/form_editor.py
@@ -11,6 +11,7 @@ from posting.widgets.variable_input import VariableInput
 class FormTable(PostingDataTable):
     BINDINGS = [
         Binding("backspace", action="remove_row", description="Remove row"),
+        Binding("space", action="toggle_row", description="Toggle row"),
     ]
 
     def on_mount(self):
@@ -18,14 +19,18 @@ class FormTable(PostingDataTable):
         self.show_header = False
         self.cursor_type = "row"
         self.zebra_stripes = True
+        self.row_disable = True
         self.add_columns("Key", "Value")
 
     def to_model(self) -> list[FormItem]:
         form_data: list[FormItem] = []
-        # TODO - handle enabled/disabled...
         for row_index in range(self.row_count):
             row = self.get_row_at(row_index)
-            form_data.append(FormItem(name=row[0], value=row[1], enabled=True))
+            form_data.append(
+                FormItem(
+                    name=row[0], value=row[1], enabled=self.is_row_enabled_at(row_index)
+                )
+            )
         return form_data
 
 
@@ -45,5 +50,7 @@ class FormEditor(Vertical):
     def to_model(self) -> list[FormItem]:
         return self.query_one(FormTable).to_model()
 
-    def replace_all_rows(self, rows: Iterable[tuple[str, str]]) -> None:
-        self.query_one(FormTable).replace_all_rows(rows)
+    def replace_all_rows(
+        self, rows: Iterable[Iterable[str]], enableStates: Iterable[bool] | None = None
+    ) -> None:
+        self.query_one(FormTable).replace_all_rows(rows, enableStates)