+++ src/posting/widgets/request/url_bar.py
@@ -10,6 +10,7 @@ from textual.events import Blur, Paste
 from textual.message import Message
 from textual.widgets import Input, Button, Label
 from textual.theme import Theme
+from textual.widgets.input import Selection
 from textual_autocomplete import DropdownItem
 from textual_autocomplete._autocomplete2 import TargetState
 from posting.config import SETTINGS
@@ -69,14 +70,6 @@ It's recommended you create a new request before pasting a curl command, to avoi
         def control(self) -> "UrlInput":
             return self.input
 
-    @dataclass
-    class Blurred(Message):
-        input: "UrlInput"
-
-        @property
-        def control(self) -> "UrlInput":
-            return self.input
-
     def on_mount(self):
         self.highlighter = VariablesAndUrlHighlighter(self)
         self.app.theme_changed_signal.subscribe(self, self.on_theme_change)
@@ -85,11 +78,8 @@ It's recommended you create a new request before pasting a curl command, to avoi
     def on_change(self, event: Input.Changed) -> None:
         self.remove_class("error")
 
-    def on_blur(self, _: Blur) -> None:
-        self.post_message(self.Blurred(self))
-
-    def watch_cursor_position(self, cursor_position: int) -> None:
-        self.post_message(self.CursorMoved(cursor_position, self.value, self))
+    def watch_selection(self, selection: Selection) -> None:
+        self.post_message(self.CursorMoved(selection.end, self.value, self))
 
     def on_theme_change(self, theme: Theme) -> None:
         super().on_theme_change(theme)
@@ -154,7 +144,7 @@ class UrlBar(Vertical):
         with Horizontal():
             yield MethodSelector(id="method-selector")
             yield UrlInput(
-                placeholder="Enter a URL or paste a curl command...",
+                placeholder="Enter a URL or paste a curl command…",
                 id="url-input",
             )
             yield Label(id="trace-markers")
@@ -183,8 +173,8 @@ class UrlBar(Vertical):
         except NoMatches:
             return
 
-    @on(UrlInput.Blurred)
-    def on_blur(self, event: UrlInput.Blurred) -> None:
+    @on(Input.Blurred)
+    def on_blur(self, event: Input.Blurred) -> None:
         try:
             self.variable_value_bar.update("")
         except NoMatches:
@@ -200,7 +190,6 @@ class UrlBar(Vertical):
         cursor_position = url_input.cursor_position
         value = url_input.value
         variable_at_cursor = get_variable_at_cursor(cursor_position, value)
-
         variables = get_variables()
         try:
             variable_bar = self.variable_value_bar